+++ libsolidity/ast/ASTUtils.h
@@ -22,8 +22,9 @@ namespace solidity::frontend
 
 class VariableDeclaration;
 
-/// Find the topmost referenced variable declaration when the given variable
+/// Find the topmost referenced constant variable declaration when the given variable
 /// declaration value is an identifier. Works only for constant variable declarations.
-VariableDeclaration const* rootVariableDeclaration(VariableDeclaration const& _varDecl);
+/// Returns nullptr if an identifier in the chain is not referencing a constant variable declaration.
+VariableDeclaration const* rootConstVariableDeclaration(VariableDeclaration const& _varDecl);
 
 }